:hammer: BE: Hotfix - Perubahan Tipe Data Harga Menjadi INT

+++ server/models/Barang.js
@@ -61,7 +61,7 @@ const Barang = db.define('Barang', {
     allowNull: true
   },
   harga_barang: {
-    type: Sequelize.REAL,
+    type: Sequelize.INTEGER,
     allowNull: true
   },
   keadaan_barang: {
